+++ server/schema/admin/res.js
@@ -1,8 +1,14 @@
 import {Admin} from '../../models'
 import bcrypt from 'bcrypt'
 import jwt from 'jsonwebtoken'
+import {_authAdmin} from '../../util'
 const resolvers = {
-  RootQuery: {},
+  RootQuery: {
+    async fetchAdmin(_, __, {loggedInUser}) {
+      _authAdmin(loggedInUser)
+      return await Admin.findById(loggedInUser.id)
+    },
+  },
   RootMutation: {
     async loginAdmin(_, {input}) {
       const admin = await Admin.findOne({where: {username: input.username}})